+++ demo/index.js
@@ -157,9 +157,32 @@
                 refresh: 'refresh9',
             });
         },
+        section10() {
+            startDemo('./deformation.js', {
+                code: 'code10',
+                preview: 'preview',
+                video: 'videos6',
+                refresh: 'refresh10',
+            });
+        },
+        section11() {
+            startDemo('./channel-split.js', {
+                code: 'code11',
+                preview: 'preview',
+                refresh: 'refresh11',
+            });
+        },
+
+        section12() {
+            startDemo('./multi-pointer.js', {
+                code: 'code12',
+                preview: 'preview',
+                refresh: 'refresh12',
+            });
+        },
     };
 
-    insertSection('section7');
+    insertSection('section11');
 
     // on load get active section from url hash and set
     const hash = window.location.hash.substring(1);